+++ maven-pmd-plugin/src/it/mpmd-168/invoker.properties
@@ -1,0 +1,1 @@
+invoker.goals = clean verify site
